Implement comment update and deletion features; add methods in PostService and PostController for updating and deleting comments, and enhance CommentResponseDTO with an ID field. Also, add functionality to like or dislike posts and retrieve likes of a post.

+++ src/main/java/com/project/Blog_Management_System/Service/UserServiceImpl.java
@@ -124,9 +124,13 @@ public class UserServiceImpl implements UserService {
                 .build();
 
         if (followDTO.getFollow()) {
-            followRepository.saveAndFlush(followEntity);
+            if (followRepository.findByFollowerAndFollowing(follower, followee).isEmpty()) {
+                followRepository.saveAndFlush(followEntity);
+            }
         } else {
-            followRepository.deleteByFollowerAndFollowing(follower, followee);
+            if (followRepository.findByFollowerAndFollowing(follower, followee).isPresent()) {
+                followRepository.deleteByFollowerAndFollowing(follower, followee);
+            }
         }
     }
